requirements.txt
bunch==1.0.1
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
httplib2==0.12.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsons==0.5.3
lml==0.0.4
MarkupSafe==1.0
numpy==1.15.2
oauth2client==4.1.3
pandas==0.23.4
Pillow==5.3.0
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
texttable==1.4.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
xlrd==1.1.0
